Allow celery and redis to be optional dependencies (closes #232)
