+++ .github/workflows/ci.yml
@@ -13,6 +13,7 @@ jobs:
     strategy:
       matrix:
         python-version: [2.7, 3.6, 3.7, 3.8, 3.9]
+        extras-required: [".", ".[redis]"]
 
     services:
       redis:
@@ -45,7 +46,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade "${{ env.pip_v }}" setuptools wheel
-        python -m pip install -e .
+        python -m pip install -e ${{ matrix.extras-required }}
         python -m pip install -r requirements-test.txt
 
     - name: Build
