+++ run-tests.sh
@@ -24,11 +24,27 @@ function deploy {
     source deactivate
   fi
 
+  if [ ! -f ~/.pyenv/versions/tracker27redis ]; then
+    pyenv virtualenv 2.7.18 tracker27redis
+    pyenv activate tracker27redis
+    pip install .[redis]
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
   # pyenv install 3.5.10
   if [ ! -f ~/.pyenv/versions/tracker35 ]; then
     pyenv virtualenv 3.5.10 tracker35
     pyenv activate tracker35
     pip install .
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
+  if [ ! -f ~/.pyenv/versions/tracker35redis ]; then
+    pyenv virtualenv 3.5.10 tracker35redis
+    pyenv activate tracker35redis
+    pip install .[redis]
     pip install -r requirements-test.txt
     source deactivate
   fi
@@ -42,11 +58,27 @@ function deploy {
     source deactivate
   fi
 
+  if [ ! -f ~/.pyenv/versions/tracker36redis ]; then
+    pyenv virtualenv 3.6.12 tracker36redis
+    pyenv activate tracker36redis
+    pip install .[redis]
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
   # pyenv install 3.7.9
   if [ ! -f ~/.pyenv/versions/tracker37 ]; then
     pyenv virtualenv 3.7.9 tracker37
     pyenv activate tracker37
     pip install .
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
+  if [ ! -f ~/.pyenv/versions/tracker37redis ]; then
+    pyenv virtualenv 3.7.9 tracker37redis
+    pyenv activate tracker37redis
+    pip install .[redis]
     pip install -r requirements-test.txt
     source deactivate
   fi
@@ -60,11 +92,27 @@ function deploy {
     source deactivate
   fi
 
+  if [ ! -f ~/.pyenv/versions/tracker38redis ]; then
+    pyenv virtualenv 3.8.6 tracker38redis
+    pyenv activate tracker38redis
+    pip install .[redis]
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
   # pyenv install 3.9.0
   if [ ! -f ~/.pyenv/versions/tracker39 ]; then
     pyenv virtualenv 3.9.0 tracker39
     pyenv activate tracker39
     pip install .
+    pip install -r requirements-test.txt
+    source deactivate
+  fi
+
+  if [ ! -f ~/.pyenv/versions/tracker39redis ]; then
+    pyenv virtualenv 3.9.0 tracker39redis
+    pyenv activate tracker39redis
+    pip install .[redis]
     pip install -r requirements-test.txt
     source deactivate
   fi
@@ -75,8 +123,16 @@ function run_tests {
   pyenv activate tracker27
   pytest -s
   source deactivate
-  
+
+  pyenv activate tracker27redis
+  pytest -s
+  source deactivate
+
   pyenv activate tracker35
+  pytest
+  source deactivate
+
+  pyenv activate tracker35redis
   pytest
   source deactivate
 
@@ -84,7 +140,15 @@ function run_tests {
   pytest
   source deactivate
 
+  pyenv activate tracker36redis
+  pytest
+  source deactivate
+
   pyenv activate tracker37
+  pytest
+  source deactivate
+
+  pyenv activate tracker37redis
   pytest
   source deactivate
 
@@ -92,18 +156,32 @@ function run_tests {
   pytest
   source deactivate
 
+  pyenv activate tracker38redis
+  pytest
+  source deactivate
+
   pyenv activate tracker39
+  pytest
+  source deactivate
+
+  pyenv activate tracker39redis
   pytest
   source deactivate
 }
 
 function refresh_deploy {
   pyenv uninstall -f tracker27
+  pyenv uninstall -f tracker27redis
   pyenv uninstall -f tracker35
+  pyenv uninstall -f tracker35redis
   pyenv uninstall -f tracker36
+  pyenv uninstall -f tracker36redis
   pyenv uninstall -f tracker37
+  pyenv uninstall -f tracker37redis
   pyenv uninstall -f tracker38
+  pyenv uninstall -f tracker38redis
   pyenv uninstall -f tracker39
+  pyenv uninstall -f tracker39redis
 }
