+++ snowplow_tracker/__init__.py
@@ -1,6 +1,12 @@
 from snowplow_tracker._version import __version__
 from snowplow_tracker.subject import Subject
-from snowplow_tracker.emitters import logger, Emitter, AsyncEmitter, CeleryEmitter, RedisEmitter
+from snowplow_tracker.emitters import logger, Emitter, AsyncEmitter
 from snowplow_tracker.self_describing_json import SelfDescribingJson
 from snowplow_tracker.tracker import Tracker
 from contracts import disable_all as disable_contracts, enable_all as enable_contracts
+
+# celery extra
+from .celery import CeleryEmitter
+
+# redis extra
+from .redis import RedisEmitter, RedisWorker
