+++ snowplow_tracker/celery/__init__.py
@@ -1,0 +1,2 @@
+from .celery_emitter import CeleryEmitter
+
