+++ snowplow_tracker/redis/__init__.py
@@ -1,0 +1,2 @@
+from .redis_emitter import RedisEmitter
+from .redis_worker import RedisWorker
